NCBC-2042: JSON.NET uses Pascal casing instead of Camel casing on Ubuntu16
Motivation
----------
The test
FtsDataMapperTests.Facets_Are_Populated_When_Result_Contains_Facets
fails on Ubuntu16 because when comparing the JSON strings, JsonConvert
is used, but it appears that on some platforms the
DefaultContractResolver is used which means the the results are Pascal
cased and not Camel cased, thus failing the test.

Modifications
-------------
 - Make the DeserilizationSettings use the
 CamelCasePropertyContractResolver instead of its default in the test,
 similar to how the SDK is configured internally.

Result
------
All search tests should succeed.

Change-Id: Ic36c05bf58e8e9706c718c871e2ea7d7f417470b
Reviewed-on: http://review.couchbase.org/112438
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>