NCBC-2040: Try to bootstrap before submitting queries
MOTIVATION
----------
When executing a query (n1ql, search or analytics) on a cluster that
does not support GCCCP and has not completed bootstrapping a client
will fail to find a node and uri to dispatch the query to.

MODIFICATIONS
-------------
- use Configuration global nodes to get query node in QueryClient,
  AnalyticsClient and SearchClient
- add hasBootstrapped field and EnsureBootstrapped method
- add semaphoreslim lock to control the _hasBootstrapped field
- add TryGetRandom extension methods to add safe way to try and get node
- add ServiceNotAvailableException and use in each service if a
  node is not available to dispatch a query to
- add integration test to verify a query can successfully be executed

RESULT
------
If bootstrapping has not completed when a query is submitted, the
client will attempt to bootstrap using a named bucket in the
configuration.

Change-Id: I69c9eb422b3eb205897f722bfc53568f2b4b3639
Reviewed-on: http://review.couchbase.org/112265
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>