NCBC-2048: Set CAS for operation results (Get, Exists)
MOTIVATION
----------
When performing operations, eg GET / INSERT, the CAS value of the
document should be returned as part of the result. This is currently
always set to 0 (zero).

MODIFICATIONS
-------------
- ensure the CAS is set using the header in OperationBase
- use the observe value's CAS value for Observe operations
- add integration tests for both Get and Exists operations

RESULT
------
When performing operations, the CAS value of the document is correctly
set. This includes standard operations eg GET, INSERT, REPLACE and
OBSERVE.

Change-Id: I74fed14aa8592ef3ca96c55cd7ad1e023111e116
Reviewed-on: http://review.couchbase.org/113197
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>