NCBC-2052: Make SDK components support Unit Testing with Mocks
Motivation
----------
As side effect of the goal to make the internals of the SDK simpler,
mocking components of the SDK was impossible. This commit makes some
changes to make mocking easier.

Modifications
-------------
 - Refactor ClusterNode into an interface IClusterNode
 - Make Couchbase.NetClient a friend assembly to DynamicProxyGenAssembly2
 - Refactor HttpClusterMap into an interface IHttpClusterMapBase
 - Make changes to unit tests to enable mocking

Result
------
Most parts of bootstrapping are now mockable; this commit stops short of
allowing mocking on all extension methods, especially for IConnection.

Change-Id: Iedd660c48368204fe05bc74eb4cb94adca194525
Reviewed-on: http://review.couchbase.org/113351
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>