NCBC-2063: Add Datastructures v2 for "SDK3"
Motivation
----------
Adds persistent implmentations of ISet, IDictionary, IList and IQueue
data structures that use Couchbase server as a backing store. The base
interfaces are all part of System.Collections.Generics, however, they
include async overloads of most of the methods for supporting the IO
involved with getting and mutating documents. In most cases,
Sub-Document API is used to limit the amout of data sent back and forth
between the client and the server; in some cases we regress back to full
document operations when the Sub-Document API doesn't support an
operation.

Modifications
-------------
 - Add IPersistentQueue, IPersistentSet, IPersistentList and
 IPersistentDictionary interfaces
 - Provide implementions of the interfaces ^
 - Add unit tests (still need some work)
 - Add extension methods for accessing via the Couchbase.ICollection interface

Result
------
The SDK now supports mechanisms for working with documents as a backing
store for common data structures.

Change-Id: I925259e053c720c06ff45a8984e2fa6b91498103
Reviewed-on: http://review.couchbase.org/113596
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>