NCBC-2088: Fix bootstrapping isssues introduced by refactoring
Motivation
----------
This is a regression caused by NCBC-2001 and perhaps NCBC-2023 for Couchbase
buckets where a hardcoded port value is 8091 instead of 11210 and because an
exception is thrown and not handled when a non-GC3P cluster is boostrapped.

Modifications
-------------
 - Use port 11210 instead of 8091 in BucketBase
 - Make call to InitializeAsync syncronous in Cluster
 - Revert to previous behaviour if GC3P cluster map call is non-success

Result
------
These are fixes to unblock the master branch; future commits will
improve or resolve work-arounds.

Change-Id: Ia5f14c199af7b95946a872be33859330df2a5cd1
Reviewed-on: http://review.couchbase.org/114102
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>