NCBC-2089: Correctly parse collection UID as hex value
MOTIVATION
----------
The server transmits collection's uid as hex when retrieving the
cluster manifest. However, this is causing the SDK to fail bootstrap
because it does not parse hex values, eg "a" as 10.

MODIFICATIONS
-------------
- add a base (16) parameter to Convert.UInt32 when converting the
  collection uid to an integer to parse hex

RESULT
------
The SDK can now successfully bootstrap if a collection with a uid
defined as a hex value is defined.

Change-Id: Ia8a39c7166c15f78466e10af99c3646b9ef2e48a
Reviewed-on: http://review.couchbase.org/114181
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>