NCBC-2092: Add array overloads to ArrayAppend, ArrayPrepend and ArrayInsert
Motivation
----------
Strips the brackets '['  and ']' from the passed array so that the
server-side API can correctly mutate the underlying document. The
support existed at the Operation level, but the flag needed to be set
for specific array types.

Modification
------------
 -  Add overloads to MutateInSpec and set the appropriate removeBrakets
 flag
 - Add overloads to MutateInSpecBuilder so that the flags will be added
 via MutateInSpec

Result
------
The brackets will be removed when an array value is passed into an
overload for ArrayAppend, ArrayPrepend or ArrayInsert so that the
brackets can be stripped.

Change-Id: Ifc0c2dd06c33af390b0490a3a9e208bfd151b2f7
Reviewed-on: http://review.couchbase.org/114260
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>