NCBC-2095: Use TryXXX when trying to set TCS result or exception
MOTIVATION
----------
A TaskCompletionSource is used to manage the returned result / exception
to the calling application when an operation is executed. The TCS can
only have one result set and will throw an exception if SetResult,
SetException, SetCancelled is called again.

This can happen if a timeout occurs with a TimeoutException returned to
the app, but the SDK may still receive a response.

MODIFICATIONS
-------------
- replace all SetResult, SetException and SetCancelled calls to TCS with
  TryXX overload

RESULT
------
If a timeout occurs, additional exceptions are not generated by trying
to set a result / exception for TaskCompletionSource that has already
been completed.

Change-Id: I1e1a01193e1a94650394137da9d14f266f8b43a0
Reviewed-on: http://review.couchbase.org/114448
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>