NCBC-1930: Decrease the default 'TcpKeepAliveTime' to 1 minute
Motivation
----------
In SDK v2 TCP Keep Alive Time was set to 2hrs; this is too long and some
enviroments (Azure for example) this may cause timeouts becuase the
environment timeout is much lower. In this patch we reduce the Keep Alive time
to 1 minute for SDK 3.0.

Modifications
-------------
 - Add TcpKeepAliveTime, TcpKeepAliveInterval and EnableTcpKeepAlives to
 ClusterOptions.
 - Update IPEndPointExtensions to take ClusterOptions and configure the
 TCP KeepAlives accordingly.
 - Update Cluster and BucketBase to pass in the ClusterOptions

Result
------
TCP Keep Alives if set (true by default) will default to 60000ms for the
initial TCP Keep Alive check and then have a duration between checks of
1000ms.

Change-Id: I2f77b459d48c9a91fd48f13c64323119a8d8943c
Reviewed-on: http://review.couchbase.org/114578
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>