NCBC-2110: Don’t reorder nodes during bootstrap
MOTIVATION
----------
The SDK currently attempts to reorder the nodes in the bucket
configuration received from the cluster based on the server list. This
is not required and can cause the SDK to fail to bootstrap.

MODIFICATIONS
-------------
- do not attempt to reorder nodes in BucketConfigExtesnsions
- remove associated tests for reordering

RESULT
------
The SDK no longer attempts to reorder nodes in bucket configurations
received from the cluster.

Change-Id: I249de8a55aeeae43a47bd4c248ea727dc7137d37
Reviewed-on: http://review.couchbase.org/115120
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>