NCBC-2122: Remove CollectionOptions from IBucket.CollectionAsync
Motivation
----------
Removes the CollectionOptions from IBucket.Collection etc. and removes
the async/await return type and naming convention; there is no IO done
here so, nothing to await on.

Modifications
-------------
 - Remove "Async" postfix and Task return type
 - Remove CollectionOptions

Result
------
Fetching a Collection is synchronous; if an error occurs because the
collection does not exist it will be handled when the collection starts
IO with the server during the first operation.

Change-Id: I03e9eb836aaaeed8ab58e20be248db75fc522ed6
Reviewed-on: http://review.couchbase.org/115366
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>