NCBC-2144: Make NMVB publish config returned in body
Motivation
----------
This commit makes the cluster map config returned by the server be
published to the config handler which will then check to see if the
config has pending changes that the client must react to. It's a
precurser to retrys.

Modifications
-------------
 - Updated CouchbaseCollection so that the ConfigContext is passed in a
 CTOR argument; when a NMVB is encountered, the memcached body is
 deserialized into a cluster map object and published to the config
 queue for processing.
 - Updated any other classes that required changes to support the CTOR
 change.

Result
------
The client will handle configs within the memcached body and publish
them to the config queue.

Change-Id: I025bd368a04003291ad344a209a3c938855dea48
Reviewed-on: http://review.couchbase.org/116089
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>