NCBC-2148: Initialize connection pool connections in parallel
MOTIVATION
----------
When a connection pool calls Initialize, it creates and adds connections
to the pool. However, this is done serially and when the pool is large
can take a long time. Creating these connections in parallel would
improve connection pool initialize times considerably when using larger
connection pools.

MODIFICATIONS
-------------
- use Parallel.For to run connection bootstrap in concurrently in
  ConnectionPool (ssl) and SharedConnectionPool (mux)

RESULT
------
When a connection pool is initialized, it bootstraps the connections in
parallel to improve startup time.

Change-Id: I76941bd3bf589c7ece276be4ba21cd621959aca3
Reviewed-on: http://review.couchbase.org/116373
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>