NCBC-2155: Consolidate Cluster constructors
MOTIVATION
----------
The Cluster object has two constructors that can either take a
ClusterOptions instance, or a username and password pair. This duplicate
logic and is open to bugs. These should be consolidated into a single
constructor and an alternative method that uses this constructor with a
username and password pair.

MODIFICATIONS
-------------
- remove Cluster constructor that had username and password parameters
- add Cluster.Connect static method that can take username and password
  as parameters

RESULT
------
There is now a single consistent constructor for the Cluster instance,
with static methods to provide convenience ways to build a cluster.

Change-Id: I6d6ff2980405adcb8fed51e07f3de3e25e54d120
Reviewed-on: http://review.couchbase.org/116385
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>