NCBC-2159: Ensure awaiter is configured in SslConnection.Send
MOTIVATION
----------
When moving from sync to async units of work, any task by default will
block the current thread. This is not ideal in some scenarios and a task
awaiter should be configured to avoid this.

MODIFICATIONS
-------------
- configure the awaiter used in SslConnection.Send when calling
  SendAsync

RESULT
------
The awaiter used in the synchronous path is correctly configured when
calling SendAsync.

Change-Id: I40919dfd101711d3a8b6a6428850d9ad4259f7ee
Reviewed-on: http://review.couchbase.org/116582
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>