NCBC-1977: Add thread.sleep to test to ensure logger is created
MOTIVATION
----------
Occasionally a logging unit test fails before the log entry is not
available in test InMemoryLogger. This can happen if the logger hasn't
finished being created and the log entry added before the assert is
executed.

MODIFICATIONS
-------------
- add thread.sleep after making log entry

RESULT
------
The extra time ensures the logger is fully created and the log entry has
been added to the InMemoryLogger before the Assert is executed.

Change-Id: I6e15b7eac7dcce63dcae1c1f54006c61758cf02f
Reviewed-on: http://review.couchbase.org/116671
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>