NCBC-2179: LookupIn does not correctly parse config returned by NMVB
Motivation
----------
The sub document API multilookup operation does not correctly handle a Not My VBucket (NMVB) response from the server.  The updated configuration from the server is not parsed correctly and the client's vbucket map is not updated.

MultiLookup<T>.GetValue attempts to interpret the beginning portion of the configuration JSON string as a body length (SDK version 2.7.13, MultiLookup.cs line 97)

This results in spikes in memory usage as an extremely large and unneed byte array is allocated, and an ArgumentException when attempting to read past the end of the response buffer.  The execption causes the response status to be changed from VBucketBelongsToAnotherServer to ClientFailure, which causes the configuration update logic to never be invoked.

Modifications
-------------
 - Bypass converting the body bytes into OperationSpecs in MultiLookup.GetValue and instead read the config later in the operation lifetime.
 - Added a unit test and test against a NMVB packet.
 - Copied ResponsePackets.cs from Couchbase.Tests and used in Unit Test
 to verify.

Result
------
The config will properly be read later and queued for config processing;
a ClientError will not be encountered.

Change-Id: I324e67bfb7a400c8488f884e3122875498972371
Reviewed-on: http://review.couchbase.org/117823
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>