NCBC-2181: Refactor exceptions and error handling to current RFC revision
Motivation
----------
Refactor exception structure from the previous iterations into the
lastest, flattened version in the RFC and adds the ErrorContext concept,
which will be integrated into error handling in a later commit.

Modifications
-------------
 - Convert older exception hierarchy to new flatter model.
 - Add ErrorContext's for storing context specific data for each
 exception
 - Update any code to use new exceptions

Result
------
This is a partial commit where the exception hierachy has been modified
per the RFC and Error Context concept added but not completely
integrated which will be done in a later commit.

Change-Id: Ibb6eb0abad0b00f32ba0fee8a35820a8b7cb1d12
Reviewed-on: http://review.couchbase.org/117997
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>