NCBC-2186: Fix regression where collection support is not detected on the server
Motivation
----------
A regression was added in an earlier commit (NCBC-2157) in which the
collection feature was not checked, thus the manifest was not loaded.
This commit fixes that bug and ensures that the collection manifest is
loaded if the server supports it.

Modifications
-------------
 - Add a property to ClusterContext ensuring that collections are
 supported.
 - Check the Cluster.SupportsCollections and load the manifest if true.

Result
------
The collection manifest will be loaded if supported, otherwise the
client will use the default scope/collection.

Change-Id: I32eab7267c749bf2b207637feae752ce6d103d58
Reviewed-on: http://review.couchbase.org/118244
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>