NCBC-2192: MultiMutation does not encode remove value correctly
MOTIVATION
----------
We incorrectly encodes a null value as the literal string `null`
in the case of a remove command.  This causes the server to
reject the operation.

MODIFICATIONS
-------------
Special case the remove command to avoid serializing null values
to JSON and instead simply leave the value empty.

RESULT
------
Executing multi-mutation removes now succeeds.

Change-Id: I70f2cdb643527134337bfeb55ae4249cd16bba0d
Reviewed-on: http://review.couchbase.org/118394
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>