NCBC-2203: make all XxxOptions parameters optional
Motivation
----------
Any parameter that is an XxxOptions should be optional; including the
argument in a method call should not be required.

Modifications
-------------
 - Make all XxxOptions arguments optional in ICluster,
 ICouchbaseCollection and implementations optional; initialize a default
 where needed.
 - Make all XxxOptions arguments in the management classses optional

Result
------
If a ctor or method has an argument of the form XxxOptions, it will be
optional and not required in which case the defaults will be used.

Change-Id: I0ecac811a4d1bcc3184ca617dbe99a8c2d8198f6
Reviewed-on: http://review.couchbase.org/118916
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>