NCBC-2205: View Service not found after failing over EP node
Motivation
----------
If an entry-point node is failed over that supports Views, an error is
returned stating that the service not found even though it should be on
another node. This patch fixes that error.

Modifications
-------------
 - Assign the bucket to the owning bucket while processing the cluster
 map.

Result
------
After failing over the EP node, View requests will be routed to another
node if one supports the feature.

Change-Id: I659f4104d1ac1e2f814ef830ac427f0de98d83e2
Reviewed-on: http://review.couchbase.org/119114
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Will Broadbelt <william.broadbelt@couchbase.com>