NCBC-2232: QueryResult is missing the QueryMetaData
This changeset adds the QueryMetaData to the QueryResult and refactors
all related properties into it.

Note that I haven't changed any of the types (i.e. guid vs string
on RequestId), only tried to keep the interface as close to spec
as possible. I'm sure there is subsequent work needed, because I
had to put the Errors back on the IQueryResult interface - hopefully
it is not internal so the users won't see it (it should be converted
into an exception instead).

I wasn't able to remove the errors because it is used on the interface
in the result extensions.

Change-Id: I9c744975205824faac493df65ba0115e79fc29f1
Reviewed-on: http://review.couchbase.org/120540
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>