NCBC-2271: Align QueryMetaData with RFC
RequestId needs to be a string and not a guid and the warning
is renamed to QueryWarning. Since Analytics was using the query
warning before, it now has its own one as well.

Also, Metrics have been renamed to QueryMetrics and analytics
(for now) got an identical copy. Needs to be cleaned up on
the analytics side later.

Change-Id: I1655fb3d425cfcfdeb6a1dbaa23f78d6c4f1da80
Reviewed-on: http://review.couchbase.org/120557
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>