NCBC-2299: Implement non-streaming fallback for N1QL queries
Motivation
----------
If a custom ITypeSerializer is provided which doesn't also implement
IStreamingTypeDeserializer, we should fallback to reading as a block
and emulate streaming on the IQueryResult API surface. This will make
implementing custom deserializers easier if streaming is required.

Modifications
-------------
Move a lot of the details of StreamQueryResult into QueryResultBase,
also cleaning up XML code comments in the process. Rename
ReadToRowsAsync to InitializeAsync to make the purpose clearer.

Add BlockQueryResult implementation to emulate streaming, and update
QueryClient to select between the two result implementations based on
capabilities of the ITypeSerializer.

Retarget StreamingQueryResultExtensions to IQueryResult, and move to the
QueryResultExtensions class.

Results
-------
Consumers can now execute N1QL queries using custom deserializers
without being required to implement IStreamingTypeDeserializer.
Previously, the SDK would always use Newtonsoft.Json to deserialize if
IStreamingTypeDeserializer was not implemented.

Change-Id: I42789e9c8612c4e708d7924ff7679b7e61c9b61f
Reviewed-on: http://review.couchbase.org/120964
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>