NCBC-2306: Add Rows property to XXXResults to match RFC
Motivation
----------
The RFC calls for a property named Rows on query, view, and analytics
results.

Modifications
-------------
Include a Rows property which returns the result again, cast to
IAsyncEnumerable, on query, view, and analtyics results.

Results
-------
While the end result is a noop, the rows are now more discoverable via
the property and the results interfaces match the RFC and other SDKs.

Change-Id: Id4d7352fc8960c78a5205dee8f0ad149417da71a
Reviewed-on: http://review.couchbase.org/121130
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>