NCBC-2308: Make ConfigHandler use Circuit Breakers
Motivation
----------
The config handler attempts to fetch a new config every few seconds, if
nodes are down and cannot except traffic, this can cause resource issues
as sockets are in a TIME_WAIT state waiting to be cleaned up. This
commit makes the SDK go through the Circuit Breaker to fetch a config so
that sockets are not opened/closed one after another.

Change-Id: I4d38f97b374881ebb58146b4709927d5be0d8a31
Reviewed-on: http://review.couchbase.org/121157
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Will Broadbelt <william.broadbelt@couchbase.com>