NCBC-2209: Add LegacyConverter and JsonConverter
Motivation
----------
Add the JsonConverter which is the default for the SDK and
LegacyConverter for backwards compatibility with SDK2. Note that
RawJsonConverterm RawStringConverter and RawBinaryConverter are added,
but not provide an implementation at this time.

Modifications
-------------
 - Add JsonTranscoder and rename DefaultTranscoder to LegacyTranscoder
 - Add RawJsonTranscoder, RawStringTranscoder and RawBinaryTranscoder
 which do not provide an implementation
 - Update the SDK to use the JsonConverter as the default converter
 - Add unit tests for JsonTranscoder ad LegacyTranscoder.
 - Update code to use JsonTranscder by registering it as a service.
 - Create a ICircuitBreaker interface for easier testing/di
 - Register CircuitBreaker as a required service.

Results
-------
The SDK now uses the JsonConverter as the default converter. If you wish
to continue using the same converter as SDK2, then update the
configuration to use the LegacyConverter.

Change-Id: I4e82fb776bc49e1a0cb851349808f9947bebaa71
Reviewed-on: http://review.couchbase.org/121184
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>