NCBC-2327: Implement lightweight DI system
Motivation
----------
Provide a DI system internal to the Couchbase SDK, with no external
dependencies, which can be used to support injection within the context
of a cluster.

Modifications
-------------
Implement internal classes for building an IServiceProvider and create
supporting tests.

Setup a collection of registered services in ClusterOptions, with a
default set of services preregistered.

Build the IServiceProvider and store it in the ClusterContext when it is
created.

Results
-------
A lightweight DI system accessible from each ClusterContext, which can
be passed down the dependency tree. It has no external dependencies.

It is also already wired to support logging injection, and as we switch
logging to use it we can eventually deprecate LogManager.

Should be usable to support other dependencies in the future, such as
transcoders, type converters, and more.

Change-Id: I660512eb5c01abcd9467acdca89946b0fa1b7e71
Reviewed-on: http://review.couchbase.org/121273
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>