NCBC-2329: Use DI system for RetryOrchestrator logging
Motivation
----------
Move towards using DI to supply dependencies all the way down the
dependency tree, cleaning up the codebase, especially for ILogger<T>.

Modifications
-------------
Add IRetryOrchestrator interface, and make RetryOrchestrator an instance
instead of a static. Accept ILogger on the constructor.

Update Cluster, CouchbaseBucket, MemcachedBucket, and BucketBase to take
an IRetryOrchestrator on the constructor.

Add RetryAsync to BucketBase to forward retry requests to
IRetryOrchestrator since it is no longer available as an extension
method.

Results
-------
RetryOrchestrator is now handled via DI and gets its logger from the DI
system instead of a static.

Change-Id: Ibd7afaf175b879f0c48d62d1ff650c4a90b695dc
Reviewed-on: http://review.couchbase.org/121357
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>