NCBC-2337: Simplify service client unit tests
Motivation
----------
Building the ClusterContext is overkill and error prone for unit tests
related to service clients.

Modifications
-------------
Create IServiceUriProvider and ServiceUriProvider to provide a more
easily mockable interface for acquiring the URIs for specific services.

Remove ClusterContext and DataMapper from HttpServiceBase. Only store
an IDataMapper on SearchClient, which actually uses it.

Update QueryClient, AnalyticsClient, and SearchClient to use an
IServiceUriProvider. For ViewClient, just remove ClusterContext as it
gets the URI from the ViewQuery.

Update all related unit tests to use the simpler approach.

Results
-------
Cleaner DI on the service clients and much simpler (and therefore less
regression prone) unit tests.

Change-Id: I3e8c3071bca8a6215391cc428c5973d76642bf3f
Reviewed-on: http://review.couchbase.org/121461
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>