NCBC-2341: Use DI for OrphanedResponseLogger logging
Motivation
----------
Step in the overall path to injecting ILogger down the dependency tree
from Cluster.

Modifications
-------------
Drop the static OrphanedResponseLoggerFactory and
NullOrphanedResponseLogger.Instance in favor of the DI system.

Change OrphanedResponseLogger to use an injected logger.

Change ClusterOptions to swap the IOrphanedResponseLogger in DI based
on the setting.

Results
-------
OrphanedResponseLogger nows uses an injected logger instead of statics.

Note: There is currently no consumer of IOrphanedResponseLogger, so it
is non-functional.

Change-Id: I6627b6bcf88bf4642980f2da4747706af13da11d
Reviewed-on: http://review.couchbase.org/121523
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>