NCBC-2345: Use DI for DNS resolver logging
Motivation
----------
Step in the overall path to injecting ILogger down the dependency tree
from Cluster.

Modifications
-------------
Rename LambdaServiceFactory to TransientServiceFactory, a more accurate
description of what it does. Add support for transient services with
automatic constructor selection.

Refactor DnsClientDnsResolver to be requested via IServiceProvider and
to accept a logger in its constructor.

Add CancellationToken to IDnsResolver for future-proofing, and enable
nullable reference types.

Change ClusterOptions to use DI infrastructure for setting a custom
DNS resolver, much like transcoders and serializers.

Results
-------
DNS resolver uses an injected logger instead of a static, and is
managed cleanly by the DI infrastructure. It is treated as a transient
instead of a singleton because it is only used briefly during bootstrap,
so there's no point in keeping it on the heap indefinitely.

When a custom DNS resolver is provided, this is treated a singleton
because a pre-existing object instance is passed to ClusterOptions.

Change-Id: Ia6ec28dce8feb25db9487651e75abd2d93ef6e2d
Reviewed-on: http://review.couchbase.org/121543
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>