NCBC-2347: Use DI for logging in query URI testers
Motivation
----------
Step in the overall path to injecting ILogger down the dependency tree
from Cluster.

Modifications
-------------
Switch to injecting the logger in the constructor of QueryUriTester and
SearchUriTester.

Enable null reference types and do a bit of log cleanup.

Results
-------
Logging for configuration is now injected rather than using statics.
No affect on the SDK for now as the types are not currently in use.

Change-Id: Ic4c4150273289166680785a2c1be7bb90eca6cd1
Reviewed-on: http://review.couchbase.org/121546
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>