NCBC-2351: Remove unused view timeout field
Motivation
----------
The _timeout field on ViewQuery is unused, there is a Timeout property
instead.

Modifications
-------------
Drop the _timeout field, and remove the unused code that places
"connection_timeout" on the view query string.

Results
-------
View query behavior is unchanged relative to previous .NET SDK 3.0 beta
releases. However, the ability to apply "connection_timeout" to view
query parameters is gone compared to .NET SDK 2.x. This appears to be
consistent, as the Java SDK 3 doesn't offer this query parameter and it
is very poorly documented.

Change-Id: I9a6e7c4b68c0c5fb5e18f41c91c4967181152baf
Reviewed-on: http://review.couchbase.org/121608
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>