NCBC-2220: Ensure that ClusterOption properties are integrated into the SDK
Motivation
----------
Wire up ClusterOptions properties so that the are the default for the
various operation refactoring and improving the structure at the same
time to reduce duplication and number of parameters.

Modifications
-------------
 - Wireup ClusterOption properties and update unit tests.
 - Remove CollectionExtensions and move methods into ClusterNode
 - Consolidate Timeout resolution if null to default in ClusterNode
 - IdleHttpConnectionTimeout throws a NotSupporteException as their is
 no mapping directly to the HttpHandlers at this time.

Results
-------
Nearly all ClusterOptions properties are correctly wired up with
defaults and can be overridden by the operations specific options. The
few that could not be at this time will be done in subsequent commits
when functionality exists.

Change-Id: I245e8380f81b8280fcad1caac00251ee3f2c197e
Reviewed-on: http://review.couchbase.org/121838
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>