NCBC-2373: Connection pooling abstraction layer
Motivation
----------
The current connection approach is a single connection, but we know we
want to add connection pooling. However, over time the connection
pooling approach may change as we learn better heuristics for improving
performance. This should be behind an abstraction layer so we can easily
and safely perform experiments.

Modifications
-------------
Move IConnection to the Couchbase.Core.IO.Connections namespace for
consistency (required touching a lot of files to add using statements).

Add IConnectionPool, ConnectionPoolFactory, and ConnectionPoolBase as
the base of the abstraction layer. Implement SingleConnectionPool to
replicate the current single-connection behavior.

Move initialization of the initial connection from ClusterNode to the
connection pool. However, some of the initialization logic is
ClusterNode specific, so add the IConnectionInitializer interface on
ClusterNode and pass this on to the ConnectionPoolFactory so that the
pool can callback when initializing new connections.

Update some of the operation implementations on ClusterNode to actually
use the connection parameter and to accept a CancellationToken.

Cleanup some of the properties on ClusterNode to be readonly and passed
in via the constructor. Update ClusterNode.SelectBucket to be async, and
to accept an IBucket instead of a bucket name. Use it to set the Owner
property rather than requiring the consumer to do both.

Update diagnostics to run against all connections in the connection pool
rather than against a single connection.

Results
-------
Behavior of the SDK is still the same, but code is a bit better
organized and concepts are less leaky. We're now in position to begin
developing an actual connection pool.

Note: SelectBucketAsync may eventually be removed from
IClusterInitializer and made part of InitializeAsync once deferred
bootstrapping is implemented.

Change-Id: I55d8a8f7ad0d3c8d47232c82256a7b537207b8a7
Reviewed-on: http://review.couchbase.org/122009
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>