NCBC-2370: Respect port numbers on connection strings
Motivation
----------
Use port numbers supplied on the connection string for bootstrapping.

Modifications
-------------
Create a new HostEndpoint structure to store hostname/port pair, with
the port being optional. Move the parsing logic for host/port pairs from
VBucketServerMapFactory to this shared structure.

Update ConnectionString to store a collection of these, already parsed,
instead of a collection of host strings. Actually include these ports,
if present, in the result of GetBootstrapUris.

Update IDnsResolver to return a list of HostEndpoints instead of a list
of Uris when resolving DNS SRV records. The scheme included in the Uri
was redundant and unused.

Update ClusterContext to use the port in CreateAndBootStrapBucketAsync.

Remove the (now unused) KvPort from ClusterOptions.

Results
-------
Ports included on the connection string are now respected and used
for bootstrapping instead of the default ports of 11210 and 11207.

Ports included in DNS SRV results are now also respected instead of
being ignored.

Change-Id: Ibf757380e29aacf587b867aaf1a591c27eea6612
Reviewed-on: http://review.couchbase.org/122078
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>