NCBC-1989: Connection Pooling
Motivation
----------
Allow each node to have a pool of connections which are shared
efficiently to prevent backups writing operations.

Modifications
-------------
Added the TPL data flows package from Microsoft.

Created DataFlowConnectionPool to use data flows to manage async queues
of operations to be written shared between multiple connections.

Removed ClusterOptions from ConnectionPoolBase, only use this in
ConnectionPoolFactory to set properties on the pool.

Added MaxKvConnections to ClusterOptions. Updated ConnectionPoolFactory
to create the legacy SingleConnectionPool only if the NumKvConnections
and MaxKvConnections values are 1, otherwise make a
DataFlowConnectionPool.

Results
-------
By default, a connection pool is used to shared the load of writing
operations. This can be disabled by the user by setting the minimum and
maximum pool size to 1.

Note: The pool does not auto-scale beyond the MinimumSize yet, this
functionality will be added later.

Change-Id: I5fabf58ac69a48ba4c9068f77b66db1a0e7730f4
Reviewed-on: http://review.couchbase.org/122139
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>