NCBC-2384: Fix SSL node lookup for K/V ops
Motivation
----------
Currently, all K/V operations fail with node not found if you are
connected via SSL. This is because nodes are indexed with the SSL port
number, but the cluster map for buckets uses the non-SSL port number.

Modifications
-------------
Add KeyEndPoints property to IClusterNode to list all endpoints (both Kv
and KvSsl) that are represented by that cluster node. And
KeyEndPointsChanged event to allow subscribers to monitor this list for
changes (which happens with delayed bootstrapping).

Implement ClusterNodeCollection to manage a collection of IClusterNode
objects, keyed by all of their KeyEndPoints, and monitor them for
changes.

Update ClusterContext to use a ClusterNodeCollection.

Upgrade to the latest version of Moq for testing, which has support for
verifying event handler registration.

Results
-------
Cluster nodes can now be found via their SSL or non-SSL K/V ports, which
allows clusters bootstrapped using SSL to function in the SDK. Also
maintains behaviors where the node can be found using it's bootstrap
endpoint, and does so while keeping a clean separation of concerns
between classes. Finally, still functions even when the NodesAdapter is
not applied until later due to delayed bootstrapping.

Change-Id: I0ddb378be9c1b22512f4dce5ab97ced4bc6e6fed
Reviewed-on: http://review.couchbase.org/122417
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>