NCBC-2387: Reduce k/v operation result heap allocations
Motivation
----------
Reduce the number of heap allocations required when processing the
result of key/value operations.

Modifications
-------------
Move the task completion source from SaslMechanismBase and ClusterNode
to OperationBase. Move the use of Read to completion on OperationBase.

Rework SocketAsyncStateExtensions to ResponseStatusExtensions.

Change ReadAsync on OperationBase to Read, and add CancellationToken to
SendAsync. Remove ReadAsync and ExtractData from IOperation altogether.

Rework ExecuteOp in ClusterNode to pass some values back to the Func
callback.

Remove unnecessary Task return type on IConnection completed callback,
it's never awaited.

Results
-------
A closure is no longer allocated on the heap in ClusterNode.ExecuteOp or
SaslMechanismBase.SendAsync to store the TaskCompletionSource for the
lambda.

Code duplication between these two methods is also reduced, as they no
longer need to manage the TaskCompletionSource, the bytes, or the call
to Read.

Reading the result data is no longer asynchronous, eliminating an await
continuation on every operation.

Calling the ExecuteOp public overloads on ClusterNode no longer
allocates a closure on the heap for the callback lambda.

Change-Id: I449a2107abe093cf918d0bc1fa77e7e27f49fdaa
Reviewed-on: http://review.couchbase.org/122434
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>